requirements.txt
# LangGraph migration dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# Keep existing dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
googlesearch-python>=1.2.3
google-search-results>=2.4.2
requests>=2.31.0
google-api-python-client>=2.100.0

# New API dependencies
tweepy>=4.14.0  # Twitter API v2

# UI Dependencies
streamlit>=1.30.0

# Testing Dependencies
pytest-playwright>=0.4.0
pytest-asyncio>=0.23.0
pytest-dotenv>=0.5.2

# Dev tooling
pre-commit>=4.0.0
detect-secrets>=1.5.0
ruff>=0.15.0
